- Renamed the example function and cpp file. - Updated the example to be able to clear the back buffer and now properly renders 1 frame ahead
