+++ TESTS/Source.cpp
@@ -1,9 +1,9 @@
 #include <Windows.h>
 
-extern void TestD3DHelpers();
+extern void D3D12HelloWorld();
 
 int APIENTRY wWinMain(_In_ HINSTANCE hInstance, _In_opt_ HINSTANCE hPrevInstance, _In_ LPWSTR lpCmdLine, _In_ int nCmdShow)
 {
-    TestD3DHelpers();
+    D3D12HelloWorld();
 
 }
